Add fcntl.h and change open() parms to O_xxx instead of integer literals

+++ src/libc/system/malloc.c
@@ -137,7 +137,7 @@ malloc(unsigned nbytes)
             return(NULL);
         }
 #endif
-        q = (NPTR)sbrk(temp*WORD);
+        q = (NPTR)sbrk((int)(temp*WORD));
         if((INT)q == -1) {
             errno = ENOMEM;
             return(NULL);
